+++ internal/provider/incident_custom_field_data_source_test.go
@@ -65,3 +65,11 @@ func testAccIncidentCustomFieldDataSourceConfig(payload customFieldDataSourceFix
 
 	return buf.String()
 }
+
+func customFieldDefault() client.CustomFieldV2 {
+	return client.CustomFieldV2{
+		Name:        "Affected Teams",
+		Description: "The teams that are affected by this incident",
+		FieldType:   client.CustomFieldV2FieldType("multi_select"),
+	}
+}